Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

See comments in spec.

It may be a good practice to start using explicit version in requirements (not build requirements) to ensure dependencies are properly updated when updating ironic. It's not needed in first versions of the package, but as it moves on and updates are required, using ">= <required version>" is recommended.

Patch-set: 3
Label: Code-Review=-1