Update patch set 1
Patch Set 1:

I think I've finally found the issue. I've added some dummy code to the Ironic test and got the following stack trace during build:

File "/builddir/build/BUILD/ironic-6.2.1.dev654/ironic/tests/unit/drivers/modules/ilo/test_common.py", 
line 45, in <module>
    import proliantutils.ilo.client as dummyclient
  File "/usr/lib/python2.7/site-packages/proliantutils/ilo/client.py", line 21, in <module>
    from proliantutils.ilo.snmp import snmp_cpqdisk_sizes as snmp
  File "/usr/lib/python2.7/site-packages/proliantutils/ilo/snmp/snmp_cpqdisk_sizes.py", line 20, in <module>
    from pysnmp import hlapi
ImportError: cannot import name hlapi

The issue is that pysnmp.hlapi does not exist in pysnmp==4.2.5, which is the version we have in our dependency repo. We need pysnmp==4.3.2 for that. 4.3.2 is the minimum version required by proliantutils [1], and this is set in upper-constraints as the higher version [2], but it's higher than the minimum required version for OpenStack [3].

Long story short, we need pysnmp 4.3.2 in the dependency repo.

[1] https://github.com/openstack/proliantutils/blob/master/requirements.txt#L6
[2] https://github.com/openstack/requirements/blob/master/upper-constraints.txt#L304
[3] https://github.com/openstack/requirements/blob/master/global-requirements.txt#L202

Patch-set: 1