Update patch set 1
Patch Set 1:

> I don't think so. Definitely not the patch in question, it only
 > affects devstack. I rather suspect proliantutils update. Maybe we
 > need to fix some dependencies there, but it's hard to guess without
 > having a local reproducer..

I've been doing some local tests, and your guess is right, we have an issue with the latest proliantutils.

We are letting proliantutils build from the latest commit, so we now have [1] and later commits, which add a dependency on sushy. python-sushy is only built on pike, so when the ocata tests try to import ilo_client in [2], they fail and ilo_client is None, causing the exception later.

So we could to several things:

1- Pin proliantutils in Ocata to the version used by Ironic in requirements.txt (2.2.0). Ideally proliantutils should be in global-requirements, so we could test this in the gate and avoid these issues.

2- Add python-sushy to Ocata, but I'm not sure if there will be some other implications.

Wdyt? I'm leaning towards option 1.

[1]- https://github.com/openstack/proliantutils/commit/100c058a09cec405fe1ab9f012699f8f4d3e9519
[2]- https://github.com/openstack/ironic/blob/stable/ocata/ironic/tests/unit/drivers/modules/ilo/test_common.py#L45

Patch-set: 1