Update patch set 1
Patch Set 1:

Let's do both?

Having sushy in Ocata won't hurt, and can simplify testing for folks with compatible hardware. It has a minimum of dependencies (six and requests?). The only problem is that it's not in g-r for Ocata.

The benefit of #2 is that it will enable us to re-sync with upstream, if e.g. we'll need a bug fix from them.

Otherwise, #1 seems to be the way to go. Note that we don't put proliantutils to g-r because nothing *depend* on it. I wonder if we should reconsider this decision.

Patch-set: 1