Bump osprofiler dependency to 1.5.0
We have found that ironic does not actually work with 1.4.0 starting
with Pike. That has never been an issue, because a newer version
was usually installed. However, it does not happen in FFU, which fails
because of it.

See https://storyboard.openstack.org/#!/story/2002900
and https://bugzilla.redhat.com/show_bug.cgi?id=1598191#c5

Change-Id: Ifd3e3a3433a17bbcc3b7150f091af02637e2b031