Update patch set 1
Patch Set 1:

Looking at the Ironic code, I see it does not set the supports_realtime option when creating a dracclient's RAIDController. This was added by [1], however it is not present on any tagged release, and we are not pinning dracclient to the latest one.

We can either pin dracclient to the latest release, or fix this in Ironic if possible. Dmitry, what do you think?

[1] - https://github.com/openstack/python-dracclient/commit/9069b1e416187c3ad293acd87a24c1ee96a9846c

Patch-set: 1
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0