Package policy as documentation
Currently a yaml file is written to /etc/ironic/policy.json, which is
invalid and causes issues for TripleO when overriding policies.

As part of the policy-json-to-yaml work, upstream has decided not to
package the default policy in /etc/ironic/ but to package an example
policy file in /usr/share/docs instead.

For upgrades where an old /etc/ironic/policy.json exists there will be
release notes with steps to manually update existing policy.

Depends-On: https://review.opendev.org/c/openstack/ironic/+/775039
Change-Id: I9f8aa9da5634c2f2294aa98eca4d8b4aaef2dd74