Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3: Workflow-1
> > > > 
> > > > Marking wip, likely will just do this in the TripleO containers instead of further complicating things.
> > > 
> > > But puppet-ironic jobs not use containers, so would also need non container fixes.
> > > https://review.opendev.org/c/openstack/puppet-openstack-integration/+/793462
> > > https://github.com/openstack/puppet-ironic/blob/1904baa3cc39d5fbf87b5ba30f36b93b3a552a90/manifests/pxe.pp#L126
> > 
> > *sigh*  It never ends.
> 
> So I was thinking about this. It is still a mistake for us to change the packaging when we're simply doing configuration and service file creation. Puppet is fully capable of doing so on it's own as a configuration management tool. Creation of an additional alias package is also just the creation of additional debt which is unnecessary given the container launch doesn't require configuration files to be generated. Realistically, we could do exactly the same with puppet and create a systemd unit so puppet modules can be tested independently. Any objections to this? If there are none, I'm just going to abandon this change and take the mentioned path forward on the puppet module so it is at least self contained and doesn't create additional dependencies outside of the critical path.

Although we can probably find examples of what you propose across openstack, the most usual (and i'd say recommended) practice is that systemd unit files are provided in rpm subpackages and if customizations are required later, options are overridden using the systemd mechanism under /etc/systemd/system or in additional config files. Also, from packaging perspective, the general recommendation is to provide a default config file that provides a minimal set of viable options so that "dnf install foo; systemctl enable foo; systemctl start foo" just works when possible (said this given the complexity of openstack services, that's not the case in many of them), even if further customization is possible.

For consistency with the rest of RDO and following the usual best practices, my recommendation would be to implement this subpackage and use puppet-ironic or triple to just install it and do the customizations needed on top of the default config file. Said this, if you think that doing the full stuff in puppet-ironic and/or tripleo, you can probably do it although i'd look for feedback from those projects.

Patch-set: 3