Update patch set 3
Patch Set 3:

(1 comment)

I think this is the better way than generating systemd service in puppet-ironic itself.

The first point is that as Alfredo mentioned this allows us to follow the generic method to use systemd services provided by packaging.

The other benefit of this is that it helps any users using rpms but not relying on puppet. Without this change they are likely to need to create their own mechanism to provide tftp server, which looks very incomplete.

Patch-set: 3
CC: Gerrit User 682 <682@270e2033-b340-4cff-9539-693957ebf0e7>