Update patch set 3
Patch Set 3:

> 
> If you feel going with puppet it's more manageable than can do that way, imo service files are included in package than it would be easier to manage the service with puppet or manual deployments, but if idea is dynamic unit file generation then with deployment tools may be better, /me not sure if ironic provides manual ways to deploy the services, with manual deploys creating system unit files would be little hacky as compared to just using unit files shipped with rpm. Currently i see ironic-api and conductor services are also packaged and puppet and manual deploys would just be setting up config and start/restart services. I haven't checked puppet-ironic code but i believe ironic-api and conductor services unit files are used from package itself and not being created by puppet.

The ironic project has bifrost being ansible based which setup services, but is driven from code. From a product support standpoint, the documentation states one thing. Undocumented new service files lead to confusion by those trying to parse documentation. "what is this service nobody has mentioned? why is it running? why can't I launch this other service?"

The argument of for the benefit of puppet doesn't hold any weight for me. It does not perform the setup for TripleO, and manual deployers ideally will just be following documentation. *plus* it is a optional service with a huge swath of hardware out there these days.

To make an analogous argument, it is not like we've created virtual packages for mysql or apache httpd. If we extend the argument I guess we would, but then who would really use them and are we doing ourselves a disservice by putting more work on our plate than is necessary to maintain something which is not directly for our product.

> Patch Set 3:
> 
> Although we can probably find examples of what you propose across openstack, the most usual (and i'd say recommended) practice is that systemd unit files are provided in rpm subpackages and if customizations are required later, options are overridden using the systemd mechanism under /etc/systemd/system or in additional config files. Also, from packaging perspective, the general recommendation is to provide a default config file that provides a minimal set of viable options so that "dnf install foo; systemctl enable foo; systemctl start foo" just works when possible (said this given the complexity of openstack services, that's not the case in many of them), even if further customization is possible.

Who will actually use this? We're not going to install a virtual package in the tftp server container. Product wise, we're not going to expect people directly run puppet anyway.

> 
> For consistency with the rest of RDO and following the usual best practices, my recommendation would be to implement this subpackage and use puppet-ironic or triple to just install it and do the customizations needed on top of the default config file. Said this, if you think that doing the full stuff in puppet-ironic and/or tripleo, you can probably do it although i'd look for feedback from those projects.

TripleO AIUI, won't be using it. https://review.opendev.org/c/openstack/tripleo-heat-templates/+/802740  I also checked with metal3 folks who use rdo packages, and they are just directly using dnsmasq because they just doing container launches where a service file is useless for them.

> Patch Set 3:
> 
> (1 comment)
> 
> I think this is the better way than generating systemd service in puppet-ironic itself.
> 
> The first point is that as Alfredo mentioned this allows us to follow the generic method to use systemd services provided by packaging.
> 
> The other benefit of this is that it helps any users using rpms but not relying on puppet. Without this change they are likely to need to create their own mechanism to provide tftp server, which looks very incomplete.

I disagree.

Creating a service file *only* for the consumption of puppet so it can self test is ridiculous. That virtual package then has to be maintained forever without actual use, because it will never be used based upon our current container definitions.

So which product are we actually trying to service? puppet-ironic as a distinct fully usable product, or TripleO using the container model?

By creating a virtual package we won't actually use. We are setting an un-ending expectation of support and consistency in that. If we have to change again for some reason at a product level, we've doubled if not tripled the touch-points required for maintenance and continuity. We go from a change in one place to a change in multiple places, because we're not able to quantify the actual value or cost. All of this maintenance has a huge amount of cost, but if we don't have a customer paying us for exactly this, and we're putting in more glue for the sake of an unknown customer, then we're doing everyone involved a disservice.

Patch-set: 3
Reviewer: Gerrit User 511 <511@270e2033-b340-4cff-9539-693957ebf0e7>