Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Creating a service file *only* for the consumption of puppet so it can self test is ridiculous. That virtual package then has to be maintained forever without actual use, because it will never be used based upon our current container definitions.
> 
> Most concerning part of implementing this only in puppet-ironic is that we no longer have any standard way to set up tftp server in RHEL9 while it is required if a user wants to use tftp in ironic.

That is not an argument *for* creating another package and having to deal with the impact of also getting that package into a pipeline where it won't be used.

Besides, it was not like this was being used at present. If you look at the container build definition[0], you'll see it is independent. No package will be useful or helpful there.

> If a user wants to set up ironic using RDO packages but not using either TripleO or puppet, they ends up in creating their own mechanism to have tftp server.

Then they have chosen to take that journey independently. They are not on an explicitly supported path. Surely they are following some sort of documentation which covers setting such up. For example, once centos stream drops, Ironic will need to update our documentation to explicitly state to setup using dnsmasq.

> If usage of dnsmasq can be a generic way to serve tftp server in CentOS9, and usage of tftp server is still valid in ironic's usecase, then it's valuable to have that standard implementation as part of RDO packages so that users can leverage what is provided by rpms from RDO repo to implement that usage.

I don't believe it will be used. If a customer or community member *does* go to use it, it may do the wrong thing in their environment as well. Ultimately it becomes additional technical debt.

Look, if RDO, as an independent project, wants to create and service more technical debt, then I'm not going to stand in the way of it. But I'm not going to be the one who creates the additional debt.


> 
> AFAIK RDO has never abandoned usage without these two tooling(TripleO or Puppet).
> 

I don't know what to tell you. TripleO AIUI only uses puppet fro configuration file generation. That seems to be the running consensus. The containers operate completely differently. See [0].

> 
> > So which product are we actually trying to service? puppet-ironic as a distinct fully usable product, or TripleO using the container model?
> 
> I'm trying to determine which would be the best for RDO as puppet-ironic. I understand your concern caused by current misalignment between upstream and downstream but IMO that doesn't justify ignoring use case of underlying items like RDO or Puppet, unless we clearly declare RDO abandons usage of any deployment unless it is managed by TripleO.

Okay, I understand, but there are aspects that likely need an internal discussion to align appropriately. Bottom line, this comes down to an issue of staffing combined with a weight of debt. We can't afford to build any more debt without there being a positive value proposition. I firmly believe creating packages and trying to create a standard for something that will not be used is just the creation of more debt.

I'm seeking guidance from management at this point, because this has become a time sink for me.

[0]: https://github.com/openstack/tripleo-heat-templates/blob/stable/wallaby/deployment/ironic/ironic-pxe-container-puppet.yaml#L125
[1]: https://review.opendev.org/c/openstack/tripleo-heat-templates/+/802740

Patch-set: 3
Reviewer: Gerrit User 511 <511@270e2033-b340-4cff-9539-693957ebf0e7>