Update patch set 3
Patch Set 3:

(1 comment)

I'm not certain why there is push back on a service unit file. IMHO unit files should *not* be managed outside of packaging except as exceptions. The fact that we do it in tripleo (or maybe in bifrost as well?) is not a best practice.  This is an operator UX improvement if dnsmasq tftp is going to be needed for ironic to function in the future, it should be provided via a package.  

Container arguments should be avoided on packaging changes because they operate differently.  While a container could use the rpm, they don't have to which leaves decisions up to the person implementing the container.  In this case the target for review should be baremetal installation.  If a user is expected to solve for dnsmasq tftp manually, that's not a good UX. Providing a simple unit+config allows the end user simply to adjust the config and start a service.  As someone who is currently trying out openstack installations via just rpms for the first time in a while, the lack of stock basic configuration files for our recommended deployment types (e.g. via apache+wsgi) is a deficiency.  If this isn't required, it could be an optional dependency or even disabled by default. But having it here makes the operators life simplier and managing configs/services manually is less than ideal.

Patch-set: 3
CC: Gerrit User 303 <303@270e2033-b340-4cff-9539-693957ebf0e7>