Update patch set 3
Patch Set 3: -Workflow

> Patch Set 3:
> 
> (1 comment)
> 
> I'm not certain why there is push back on a service unit file. IMHO unit files should *not* be managed outside of packaging except as exceptions. The fact that we do it in tripleo (or maybe in bifrost as well?) is not a best practice.  This is an operator UX improvement if dnsmasq tftp is going to be needed for ironic to function in the future, it should be provided via a package.  

So by logical progression, we should also have an openstack-ironic-apache-httpd service as well?

Which puts in place which our products won't actually use, right?

This comes down to a matter of business value. I don't see business value in creating virtual packages for upstream testing of puppet-ironic. I see confusion and a need for additional documentation which our downstream customers shouldn't be using anyway.

And progressing down the container path means the unit files will never be used.

> 
> Container arguments should be avoided on packaging changes because they operate differently.  While a container could use the rpm, they don't have to which leaves decisions up to the person implementing the container.  In this case the target for review should be baremetal installation.  If a user is expected to solve for dnsmasq tftp manually, that's not a good UX. Providing a simple unit+config allows the end user simply to adjust the config and start a service.  As someone who is currently trying out openstack installations via just rpms for the first time in a while, the lack of stock basic configuration files for our recommended deployment types (e.g. via apache+wsgi) is a deficiency.  If this isn't required, it could be an optional dependency or even disabled by default. But having it here makes the operators life simplier and managing configs/services manually is less than ideal.


Who is the actual user we're speaking of because from a product and business value point of view, the problem is already solved. Which means this discussion is about solving it for someone who is off of the larger project upstream and downstream supported paths.

I guess I can agree in having it would make a portion of operator's lives easier if they are not following published upstream documentation and choosing to go it alone and blindly. These same operators are trying things out manually, and if there is not explicit documentation saying "use xyz packages" then we simply are creating more debt without value.

This would be an entirely different conversation if we had plans to use the service files and embedded configuration for a dnsmasq instance hosting tftp in the future, but as far as I'm aware we do not have plans to do such. Which brings us entirely back to the beginning, we're talking about doing this just so we can have puppet-ironic install a package so we can test puppet-ironic perform an independent integration test of it's generated configuration, which we'll eventually move away from anyway.

Removing WF-1, but holding any further action until I have clarification from the business. (In other words, others are free to pick this up if you *really* want or need it.)

Patch-set: 3
Reviewer: Gerrit User 511 <511@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0