Update patch set 3
Patch Set 3:

> 
> So by logical progression, we should also have an openstack-ironic-apache-httpd service as well?
> 

We don't need to ship the http service because that's already shipped via httpd. We could ship an http config similar to keystone.

[root@overcloud-controller-0 ~]# rpm -ql openstack-keystone | grep wsgi
/usr/bin/keystone-wsgi-admin
/usr/bin/keystone-wsgi-public
/usr/share/keystone/wsgi-keystone.conf

Correct me if i'm wrong here but my understanding is that we need something to provide tftp for ironic to function correct?  If there is not already a service readily available (e.g. package tftp) that we can reuse, then openstack-ironic should provide a service that handles this functionality. You can still depend on dnsmasq, but having a simple config+service unit that can be referenced would be a best practice.

> Which puts in place which our products won't actually use, right?
> 

This is an upstream change, discussion of products is not relevant here.

> This comes down to a matter of business value. I don't see business value in creating virtual packages for upstream testing of puppet-ironic. I see confusion and a need for additional documentation which our downstream customers shouldn't be using anyway.
> 

This is an upstream change, the business value may not be the driving factor.

> And progressing down the container path means the unit files will never be used.
> 
> 
> Who is the actual user we're speaking of because from a product and business value point of view, the problem is already solved. Which means this discussion is about solving it for someone who is off of the larger project upstream and downstream supported paths.
> 

Consumer of RDO packages to deploy OpenStack.  In this instance, someone who needs to configure ironic on baremetal and wants to have it work.

> I guess I can agree in having it would make a portion of operator's lives easier if they are not following published upstream documentation and choosing to go it alone and blindly. These same operators are trying things out manually, and if there is not explicit documentation saying "use xyz packages" then we simply are creating more debt without value.
> 

The upstream documentation for installation on CentOS could be updated to include using this package and leverage the provided service.

> This would be an entirely different conversation if we had plans to use the service files and embedded configuration for a dnsmasq instance hosting tftp in the future, but as far as I'm aware we do not have plans to do such. Which brings us entirely back to the beginning, we're talking about doing this just so we can have puppet-ironic install a package so we can test puppet-ironic perform an independent integration test of it's generated configuration, which we'll eventually move away from anyway.

Puppet-ironic should not be managing a systemd unit file. It should manage configuration files and service state.  Something needs to provide the service (and ideally a default config) and since ironic is the consumer, the most logical place is to put it as a subpackage here for users who which to use this configuration.  From a TripleO standpoint, the containers could use this sub package and launch the tftp server with the provide configuration which would bring it more in line with every other service we have.  The fact that we're manually doing dnsmasq stuff for this is actually an exception to how every other service gets configured (puppet for configs, containers launch a server process using these configs).

> 
> Removing WF-1, but holding any further action until I have clarification from the business. (In other words, others are free to pick this up if you *really* want or need it.)

We do want this because it's a best practice.

Patch-set: 3