Update patch set 4
Patch Set 4:

(12 comments)

> Patch Set 4:
> 
> Now I understand this is a clone of openstack-ironic-inspector-dnsmasq.service[1]. Because currently inspector-dnsmasq has tftp server enabled by default[2] I think this would conflict with that existing dnsmasq. We need to understand we can launch two dsnamasq services here at the same time(*1). I checked the behavior of dnsmasq and seems it doesn't complain duplicate binding, though.
> 
> [1]
> https://github.com/rdo-packages/ironic-inspector-distgit/blob/f373d2f19baf64ad4ada41c8cf732f0779f8416e/openstack-ironic-inspector-dnsmasq.service#L7
> 
> [2] 
> https://github.com/rdo-packages/ironic-inspector-distgit/blob/f373d2f19baf64ad4ada41c8cf732f0779f8416e/dnsmasq.conf#L4
> 
> One option is to use that existing dnsmasq instead of adding this new service(*2). It would not be ideal that ironic depends on a service from ironic-inspector but it can avoid effectively duplicate services.
> 

My impression is the inspector dnsmasq is often used separately, and in fact has been used seprately as it services an entirely non-core use case of discovering hardware not enrolled in Ironic.

So to be quite explicit. Inspector is not required. Inspector supports an entirely separate use case. I'm not even sure that file is correct, and may just be more technical debt.

From a product perspective, I'm not even sure that is getting used that way. Examining the template config, https://github.com/openstack/puppet-ironic/blob/master/templates/inspector_dnsmasq_tftp.erb, it *does* look like it is not maintained configuration, which means the setting in the ironic-inspector package is just debt/bad configuration. Also the stock config there is entirely non-functional for its integrated use case. *shrug*

> 
> Side notes
> 
> (*1) The same question would be applied to TripleO which use tftp.in for ironic_pxe_tftp and dnsmasq for ironic-inspector-dnsmasq. If we replace tftp.in by dnsmasq then we have two dnsmasq service listening on provisioning interface(one with interface=br-ctlplane and the other with listen-address=192.168.24.1) 

I think we need to edit the ironic-inspector-distgit to remove the tftp setting, at least based upon the template which overwrites the file.

> 
> (*2)
> Looking at git history, it seems puppet-ironic initially relied on ironic-inspector-dnsmasq for TFTP server but it disabled the feature when the generic ironic::pxe was introduced. 
> 
> https://github.com/openstack/puppet-ironic/commit/0376a5f55d8741139eaba9407e3863c7aabca4fc#diff-8fe3392b05f5ede5475ece31ce2bbd7047ff4a8f4bcf7916a7dcdfe4fa8eb8bc
> 
> I have proposed the change to enable tftp server of ironic-inspector-dnsmasq for you reference.
>  https://review.opendev.org/c/openstack/puppet-ironic/+/806276

That is more operational debt *and* inspector is an entirely optional component.

Patch-set: 4
Reviewer: Gerrit User 511 <511@270e2033-b340-4cff-9539-693957ebf0e7>