Update patch set 4
Patch Set 4: -Code-Review

Thanks Julia,

> So to be quite explicit. Inspector is not required. Inspector supports an entirely separate use case. I'm not even sure that file is correct, and may just be more technical debt.

OK so this is the critical reason we should have it in part of ironic.

It's better we update the dnsmasq.conf in ironic-inspector which now recommends enable-tftp but that is not a blocker now and can be done later as a follow up.

Patch-set: 4
Reviewer: Gerrit User 682 <682@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0