Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > Looks good, but I think the subpackage should also be responsible for creating the /tftpboot directory
> > > 
> > > /tftpboot is created by syslinux-tftpboot package. I'm not sure if this package is always installed in ironic pxe environments or it may be needed in cases where it's not installed.
> > 
> > syslinux-tftpboot is *entirely* optional at this point. At least from a project technical standpoint.
> 
> In fact, syslinux is not a dependency (nor should it because multiarch reasons), so perhaps we should make a folder...

A further thought. Perhaps the service should just make the folder it is configured for? It already places the required template file on startup, so it doesn't seem too unreasonable besides generally we would want the folder, wherever it is configured to ideally have more restricted permissions, but that would restrict it to the running service anyway.

Patch-set: 5
Reviewer: Gerrit User 511 <511@270e2033-b340-4cff-9539-693957ebf0e7>