Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> (2 comments)
> 
> I think we could re-use the openstack-ironic-inspector-dnsmasq.service, re-name it to just openstack-ironic-dnsmasq.service and update the puppet module to add the configuration to enable the TFTP service and/or the DHCP service for inspector? i.e just some conditions in [1] and [2]? (and renaming the tamplates to dnsmasq_http.erb, dnsmasq_tftp.erb)
> 

I disagree given inspector is not required for Ironic to run. It is not explicitly needed to deploy a node. These services can also be deployed on separate node if someone is using the RPM packages and self configuring. For what it is worth, this also came up on a puppet-ironic change and Takashi agreed upon learning ironic-inspector is not required, that we should not try to re-use the ironic-inspector provided dnsmasq deployment.

Then I also have to ask why would we want to go down that path? From my point of view, this change wouldn't be used in the actual product, and any inter-twined changes in upstream tools to support a midstream packages would just needlessly further complicate what is already kind of redundant, in that Ironic already knows how to handle pxe/ipxe for managed introspection, leaving the remaining use case "hardware discovery". Realistically, downstream product wise, the actual change required is one alrady proposed against tripleo-heat-templates.

> In puppet-ironic we can split out the "dnsmasq" config from manifests/inspector.pp to an individual manifest that is included by both manifests/inspector.pp and whatever is used to deploy ironic?
> 

Honestly, In the long term, I would prefer we split out dnsmasq completely from puppet-ironic given the amount of complexity and redundant and even incorrect configuration which can result. I think it would just be simpler moving towards a model of just one under some sort of unified management, but that is also kind of aside from packaging changes so we can continue to test puppet-ironic independent of anything else. Downstream, if we drop unknown hardware discovery... *or* if we enable ironic to *have* a configuration which enables it through an alternative template, then it would be entirely possible to maintain as long as there was something providing IP addressing to the inspecting nodes in a non-neutron managed introspection network use case.

> 
> And yes, templates/inspector_dnsmasq_[http|tftp].erb in puppet-ironic writes /var/lib/config-data/puppet-generated/ironic_inspector/etc/ironic-inspector/dnsmasq.conf in current tripleo.
> 
> [1] https://github.com/openstack/puppet-ironic/blob/master/templates/inspector_dnsmasq_http.erb
> [2] https://github.com/openstack/puppet-ironic/blob/master/templates/inspector_dnsmasq_tftp.erb

Patch-set: 7
Reviewer: Gerrit User 511 <511@270e2033-b340-4cff-9539-693957ebf0e7>